requirements.txt
streamlit
faiss-cpu
sentence-transformers
PyPDF2
transformers
huggingface_hub
requests
python-dotenv